Fix installed UE4Editor.modules file being deleted when packaging a plugin from the installed engine.
#rb none
#jira UE-65189

#ROBOMERGE-SOURCE: CL 4480784 in //UE4/Release-4.21/...
#ROBOMERGE-BOT: RELEASE (Release-4.21 -> Release-Staging-4.21)

[CL 4480793 by ben marsh in Staging-4.21 branch]

+++ Engine/Source/Programs/UnrealBuildTool/Configuration/UEBuildTarget.cs
@@ -1701,7 +1701,7 @@ namespace UnrealBuildTool
 				foreach (FileReference VersionManifestFile in FileReferenceToModuleManifestPairs.Select(x => x.Key))
 				{
 					// Make sure the file (and directory) exists before trying to delete it
-					if(FileReference.Exists(VersionManifestFile))
+					if(FileReference.Exists(VersionManifestFile) && !IsFileInstalled(VersionManifestFile))
 					{
 						FileReference.Delete(VersionManifestFile);
 					}
